+++ TimberCommons/CommonUIPatches/ManufactoryInventoryFragmentInitializeFragmentPatch.cs
@@ -10,9 +10,10 @@ using UnityEngine.UIElements;
 using ProgressBar = Timberborn.CoreUI.ProgressBar;
 
 // ReSharper disable InconsistentNaming
-namespace IgorZ.TimberCommons.IrrigationSystemUI {
+namespace IgorZ.TimberCommons.CommonUIPatches {
 
-/// <summary>Adds "supply left" element to the manufactory UI fragment.</summary>
+/// <summary>Harmony patch that adds "supply left" element to the manufactory UI fragment.</summary>
+/// <remarks>It only adds an element, but the actual handling is not made here.</remarks>
 /// <seealso cref="ManufactoryInventoryFragmentUpdateFragmentPatch"/>
 [HarmonyPatch(typeof(ManufactoryInventoryFragment), nameof(ManufactoryInventoryFragment.InitializeFragment))]
 static class ManufactoryInventoryFragmentInitializeFragmentPatch {