+++ TimberCommons/CommonUIPatches/ManufactoryInventoryFragmentUpdateFragmentPatch.cs
@@ -11,9 +11,9 @@ using UnityEngine.UIElements;
 using ProgressBar = Timberborn.CoreUI.ProgressBar;
 
 // ReSharper disable InconsistentNaming
-namespace IgorZ.TimberCommons.IrrigationSystemUI {
+namespace IgorZ.TimberCommons.CommonUIPatches {
 
-/// <summary>Handles "supply left" element in the manufactory UI fragment.</summary>
+/// <summary>Harmony patch to display "supply left" element in the manufactory UI fragment.</summary>
 /// <seealso cref="ManufactoryInventoryFragmentInitializeFragmentPatch"/>
 [HarmonyPatch(typeof(ManufactoryInventoryFragment), nameof(ManufactoryInventoryFragment.UpdateFragment))]
 static class ManufactoryInventoryFragmentUpdateFragmentPatch {