+++ TimberCommons/WaterService/Configurator.cs
@@ -2,8 +2,12 @@
 // Author: igor.zavoychinskiy@gmail.com
 // License: Public Domain
 
+using System;
+using System.Collections.Generic;
 using Bindito.Core;
+using IgorZ.TimberCommons.Common;
 using IgorZ.TimberDev.Logging;
+using IgorZ.TimberDev.Utils;
 using TimberApi.ConfiguratorSystem;
 using TimberApi.SceneSystem;
 
@@ -13,6 +17,13 @@ namespace IgorZ.TimberCommons.WaterService {
 // ReSharper disable once UnusedType.Global
 sealed class Configurator : IConfigurator {
   public void Configure(IContainerDefinition containerDefinition) {
+
+    var patches = new List<Type> { typeof(SoilMoistureSimulatorGetUpdatedMoisturePatch) };
+    if (Features.OverrideDesertLevelsForWaterTowers) {
+      patches.Add(typeof(SoilMoistureMapUpdateDesertIntensityPatch));
+    }
+    HarmonyPatcher.PatchRepeated(GetType().AssemblyQualifiedName, patches.ToArray());
+
     containerDefinition.Bind<DirectWaterServiceAccessor>().AsSingleton();
     containerDefinition.Bind<DirectSoilMoistureSystemAccessor>().AsSingleton();
     containerDefinition.Bind<ThreadedLogsRecorder>().AsSingleton();